refactor: use new block-broker interface
