--- packages/helia/src/block-providers/index.ts
@@ -1,2 +1,0 @@
-export { BitswapBlockProvider } from './bitswap-block-provider.js'
-export { TrustlessGatewayBlockProvider } from './trustless-gateway-block-provider.js'
